Excel funcionando en el login

+++ src/main/java/com/hyg/proyecto/model/Producto.java
@@ -6,34 +6,31 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-
 @Entity
-@Table(name="Producto")
+@Table(name = "Producto")
 public class Producto {
     @Id
-    @GeneratedValue(strategy =GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private String NombreP;
     private int cantidadP;
-    private double precio;
+    private int precio;
 
     public Producto() {
-       
+
         this.id = 0;
         this.NombreP = "";
         this.cantidadP = 0;
-        this.precio = 0.0;
+        this.precio = 0;
     }
 
-   
-    public Producto(int id, String NombreP, int cantidadP, double precio) {
-       
+    public Producto(int id, String NombreP, int cantidadP, int precio) {
+
         this.id = id;
         this.NombreP = NombreP;
         this.cantidadP = cantidadP;
         this.precio = precio;
     }
-
 
     public int getId() {
         return id;
@@ -59,11 +56,11 @@ public class Producto {
         this.cantidadP = cantidadP;
     }
 
-    public double getPrecio() {
+    public int getPrecio() {
         return precio;
     }
 
-    public void setPrecio(double precio) {
+    public void setPrecio(int precio) {
         this.precio = precio;
     }
 }
